requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
langchain==0.1.11
langchain-community==0.0.27
langchain-groq==0.0.1
sentence-transformers==2.5.1
faiss-cpu==1.8.0
torch==2.2.1
transformers==4.38.2
numpy==1.26.4
werkzeug==3.0.1
gunicorn==21.2.0 